ai-server/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
timm
pillow
numpy
opencv-python-headless
ttach
grad-cam